PFP-6935 Unngår å eksponere KravgrunnlagAggregate-entitet + ytelsesforbedring (indeks) + stramme opp kode som henter fakta (antar grunnlag finnes) + rydde i BehandlingTjeneste, FaktaFeilutbetalingTjeneste
