+++ domenetjenester/behandling/src/test/java/no/nav/foreldrepenger/tilbakekreving/behandling/impl/BehandlingTjenesteImplTest.java
@@ -3,7 +3,6 @@ package no.nav.foreldrepenger.tilbakekreving.behandling.impl;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.when;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.Optional;
 import java.util.UUID;
@@ -14,7 +13,6 @@ import org.junit.Test;
 import com.google.common.collect.Lists;
 
 import no.nav.foreldrepenger.tilbakekreving.FellesTestOppsett;
-import no.nav.foreldrepenger.tilbakekreving.behandling.modell.BehandlingFeilutbetalingFakta;
 import no.nav.foreldrepenger.tilbakekreving.behandlingslager.behandling.Behandling;
 import no.nav.foreldrepenger.tilbakekreving.behandlingslager.behandling.BehandlingResultatType;
 import no.nav.foreldrepenger.tilbakekreving.behandlingslager.behandling.BehandlingType;
@@ -31,10 +29,6 @@ import no.nav.foreldrepenger.tilbakekreving.fpsak.klient.dto.Behandlingsresultat
 import no.nav.foreldrepenger.tilbakekreving.fpsak.klient.dto.BehandlingÅrsakDto;
 import no.nav.foreldrepenger.tilbakekreving.fpsak.klient.dto.EksternBehandlingsinfoDto;
 import no.nav.foreldrepenger.tilbakekreving.fpsak.klient.dto.TilbakekrevingValgDto;
-import no.nav.foreldrepenger.tilbakekreving.grunnlag.Kravgrunnlag431;
-import no.nav.foreldrepenger.tilbakekreving.grunnlag.KravgrunnlagMock;
-import no.nav.foreldrepenger.tilbakekreving.grunnlag.KravgrunnlagMockUtil;
-import no.nav.foreldrepenger.tilbakekreving.grunnlag.kodeverk.KlasseType;
 
 public class BehandlingTjenesteImplTest extends FellesTestOppsett {
 
@@ -44,112 +38,6 @@ public class BehandlingTjenesteImplTest extends FellesTestOppsett {
     public void setup() {
         when(mockFpsakKlient.hentTilbakekrevingValg(eksternBehandlingUuid)).thenReturn(Optional.of(new TilbakekrevingValgDto(VidereBehandling.TILBAKEKREV_I_INFOTRYGD)));
         when(mockFpsakKlient.hentBehandling(eksternBehandlingUuid)).thenReturn(Optional.of(lagEksternBehandlingsInfo()));
-    }
-
-    @Test
-    public void skalReturnereTomFeilutbetalingFaktaNårGrunnlagIkkeFinnes() {
-        Optional<BehandlingFeilutbetalingFakta> feilutbetalingFakta = behandlingTjeneste.hentBehandlingFeilutbetalingFakta(internBehandlingId);
-        assertThat(feilutbetalingFakta).isEmpty();
-    }
-
-    @Test
-    public void skal_hente_feilutbetalingfakta_når_varselBeløp_ikke_finnes() {
-        KravgrunnlagMock mockMedFeilPostering = lagKravgrunnlag(FOM, TOM, KlasseType.FEIL, BigDecimal.valueOf(10000), BigDecimal.ZERO);
-        KravgrunnlagMock mockMedYtelPostering = lagKravgrunnlag(FOM, TOM, KlasseType.YTEL, BigDecimal.ZERO, BigDecimal.valueOf(10000));
-
-        Kravgrunnlag431 kravgrunnlag431 = KravgrunnlagMockUtil.lagMockObject(Lists.newArrayList(mockMedFeilPostering, mockMedYtelPostering));
-
-        grunnlagRepository.lagre(internBehandlingId,kravgrunnlag431);
-
-        Optional<BehandlingFeilutbetalingFakta> feilutbetalingFakta = behandlingTjeneste.hentBehandlingFeilutbetalingFakta(internBehandlingId);
-        assertThat(feilutbetalingFakta).isNotEmpty();
-        BehandlingFeilutbetalingFakta fakta = feilutbetalingFakta.get();
-        fellesFaktaResponsSjekk(fakta);
-        assertThat(fakta.getTidligereVarseltBeløp()).isNull();
-    }
-
-    @Test
-    public void skalHenteFeilutbetalingFaktaMedEnkelPeriode() {
-        KravgrunnlagMock mockMedFeilPostering = lagKravgrunnlag(FOM, TOM, KlasseType.FEIL, BigDecimal.valueOf(10000), BigDecimal.ZERO);
-        KravgrunnlagMock mockMedYtelPostering = lagKravgrunnlag(FOM, TOM, KlasseType.YTEL, BigDecimal.ZERO, BigDecimal.valueOf(10000));
-
-        Kravgrunnlag431 kravgrunnlag431 = KravgrunnlagMockUtil.lagMockObject(Lists.newArrayList(mockMedFeilPostering, mockMedYtelPostering));
-
-        grunnlagRepository.lagre(internBehandlingId,kravgrunnlag431);
-        varselRepository.lagre(internBehandlingId,"hello",23000l);
-        Optional<BehandlingFeilutbetalingFakta> feilutbetalingFakta = behandlingTjeneste.hentBehandlingFeilutbetalingFakta(internBehandlingId);
-
-        assertThat(feilutbetalingFakta).isNotEmpty();
-        BehandlingFeilutbetalingFakta fakta = feilutbetalingFakta.get();
-        fellesFaktaResponsSjekk(fakta);
-        assertThat(fakta.getAktuellFeilUtbetaltBeløp()).isEqualByComparingTo(BigDecimal.valueOf(10000));
-        assertThat(fakta.getPerioder().get(0).getBelop()).isEqualByComparingTo(BigDecimal.valueOf(10000));
-        assertThat(fakta.getTidligereVarseltBeløp()).isEqualByComparingTo(23000l);
-        assertThat(fakta.getPerioder().get(0).getFom()).isEqualTo(FOM);
-        assertThat(fakta.getPerioder().get(0).getTom()).isEqualTo(TOM);
-    }
-
-    @Test
-    public void skalHenteFeilutbetalingFaktaMedFlerePerioder() {
-        KravgrunnlagMock mockMedFeilPostering = lagKravgrunnlag(FOM, LocalDate.of(2016, 03, 31), KlasseType.FEIL,
-            BigDecimal.valueOf(10000), BigDecimal.ZERO);
-        KravgrunnlagMock mockMedFeilPostering2 = lagKravgrunnlag(LocalDate.of(2016, 04, 01), LocalDate.of(2016, 04, 15),
-            KlasseType.FEIL, BigDecimal.valueOf(12000), BigDecimal.ZERO);
-        KravgrunnlagMock mockMedFeilPostering3 = lagKravgrunnlag(LocalDate.of(2016, 04, 22), TOM,
-            KlasseType.FEIL, BigDecimal.valueOf(15000), BigDecimal.ZERO);
-        KravgrunnlagMock mockMedYtelPostering = lagKravgrunnlag(FOM, TOM, KlasseType.YTEL, BigDecimal.ZERO, BigDecimal.valueOf(37000));
-
-        Kravgrunnlag431 kravgrunnlag431 = KravgrunnlagMockUtil.lagMockObject(Lists.newArrayList(mockMedFeilPostering, mockMedFeilPostering2,
-            mockMedFeilPostering3, mockMedYtelPostering));
-
-        grunnlagRepository.lagre(internBehandlingId,kravgrunnlag431);
-        varselRepository.lagre(internBehandlingId,"hello",23000l);
-
-        Optional<BehandlingFeilutbetalingFakta> feilutbetalingFakta = behandlingTjeneste.hentBehandlingFeilutbetalingFakta(internBehandlingId);
-
-        assertThat(feilutbetalingFakta).isNotEmpty();
-        BehandlingFeilutbetalingFakta fakta = feilutbetalingFakta.get();
-        fellesFaktaResponsSjekk(fakta);
-        assertThat(fakta.getAktuellFeilUtbetaltBeløp()).isEqualByComparingTo(BigDecimal.valueOf(37000));
-        assertThat(fakta.getTidligereVarseltBeløp()).isEqualByComparingTo(23000l);
-        assertThat(fakta.getPerioder().size()).isEqualTo(2);
-        assertThat(fakta.getPerioder().get(0).getFom()).isEqualTo(FOM);
-        assertThat(fakta.getPerioder().get(0).getTom()).isEqualTo(LocalDate.of(2016, 04, 15));
-        assertThat(fakta.getPerioder().get(0).getBelop()).isEqualTo(BigDecimal.valueOf(22000));
-
-        assertThat(fakta.getPerioder().get(1).getTom()).isEqualTo(TOM);
-        assertThat(fakta.getPerioder().get(1).getFom()).isEqualTo(LocalDate.of(2016, 04, 22));
-        assertThat(fakta.getPerioder().get(1).getBelop()).isEqualTo(BigDecimal.valueOf(15000));
-    }
-
-    @Test
-    public void skalHenteFeilutbetalingFaktaMedFlerePerioderOgSisteDagIHelgen() {
-        KravgrunnlagMock mockMedFeilPostering = lagKravgrunnlag(FOM, LocalDate.of(2016, 03, 26), KlasseType.FEIL,
-            BigDecimal.valueOf(10000), BigDecimal.ZERO);
-        KravgrunnlagMock mockMedFeilPostering2 = lagKravgrunnlag(LocalDate.of(2016, 03, 28), LocalDate.of(2016, 04, 15),
-            KlasseType.FEIL, BigDecimal.valueOf(12000), BigDecimal.ZERO);
-        KravgrunnlagMock mockMedFeilPostering3 = lagKravgrunnlag(LocalDate.of(2016, 04, 19), TOM,
-            KlasseType.FEIL, BigDecimal.valueOf(15000), BigDecimal.ZERO);
-        KravgrunnlagMock mockMedYtelPostering = lagKravgrunnlag(FOM, TOM, KlasseType.YTEL, BigDecimal.ZERO, BigDecimal.valueOf(37000));
-
-        Kravgrunnlag431 kravgrunnlag431 = KravgrunnlagMockUtil.lagMockObject(Lists.newArrayList(mockMedFeilPostering, mockMedFeilPostering2,
-            mockMedFeilPostering3, mockMedYtelPostering));
-
-        grunnlagRepository.lagre(internBehandlingId,kravgrunnlag431);
-        Optional<BehandlingFeilutbetalingFakta> feilutbetalingFakta = behandlingTjeneste.hentBehandlingFeilutbetalingFakta(internBehandlingId);
-
-        assertThat(feilutbetalingFakta).isNotEmpty();
-        BehandlingFeilutbetalingFakta fakta = feilutbetalingFakta.get();
-        fellesFaktaResponsSjekk(fakta);
-        assertThat(fakta.getAktuellFeilUtbetaltBeløp()).isEqualByComparingTo(BigDecimal.valueOf(37000));
-        assertThat(fakta.getPerioder().size()).isEqualTo(2);
-        assertThat(fakta.getPerioder().get(0).getFom()).isEqualTo(FOM);
-        assertThat(fakta.getPerioder().get(0).getTom()).isEqualTo(LocalDate.of(2016, 04, 15));
-        assertThat(fakta.getPerioder().get(0).getBelop()).isEqualTo(BigDecimal.valueOf(22000));
-
-        assertThat(fakta.getPerioder().get(1).getTom()).isEqualTo(TOM);
-        assertThat(fakta.getPerioder().get(1).getFom()).isEqualTo(LocalDate.of(2016, 04, 19));
-        assertThat(fakta.getPerioder().get(1).getBelop()).isEqualTo(BigDecimal.valueOf(15000));
     }
 
     @Test
@@ -267,22 +155,6 @@ public class BehandlingTjenesteImplTest extends FellesTestOppsett {
         behandling.avsluttBehandling();
         BehandlingLås behandlingLås = repoProvider.getBehandlingRepository().taSkriveLås(behandling);
         behandlingRepository.lagre(behandling, behandlingLås);
-    }
-
-
-    private KravgrunnlagMock lagKravgrunnlag(LocalDate fom, LocalDate tom, KlasseType klasseType, BigDecimal nyBeløp, BigDecimal tilbakeBeløp) {
-        return new KravgrunnlagMock(fom, tom, klasseType, nyBeløp, tilbakeBeløp);
-    }
-
-    private void fellesFaktaResponsSjekk(BehandlingFeilutbetalingFakta fakta) {
-        assertThat(fakta.getTotalPeriodeFom()).isEqualTo(FOM);
-        assertThat(fakta.getTotalPeriodeTom()).isEqualTo(TOM);
-        assertThat(fakta.getPerioder()).isNotEmpty();
-        assertThat(fakta.getDatoForRevurderingsvedtak()).isEqualTo(NOW);
-        assertThat(fakta.getTilbakekrevingValg().getVidereBehandling()).isEqualToComparingFieldByField(VidereBehandling.TILBAKEKREV_I_INFOTRYGD);
-        assertThat(fakta.getBehandlingsresultat().getType()).isEqualByComparingTo(BehandlingResultatType.OPPHØR);
-        assertThat(fakta.getBehandlingsresultat().getKonsekvenserForYtelsen()).contains(KonsekvensForYtelsen.ENDRING_I_BEREGNING);
-        assertThat(fakta.getBehandlingÅrsaker().size()).isEqualTo(1);
     }
 
     private void fellesBehandlingAssert(Long behandlingId,boolean manueltOpprettet) {
