+++ migreringer/src/main/resources/db/migration/defaultDS/1.0/V1.1_139__PFP-6935-index-kravgrunnlag.sql
@@ -1,0 +1,2 @@
+--index brukes for oppslag når det mottas et kravgrunnlag, se KravgrunnlagRepository
+create index IDX_KRAV_GRUNNLAG_431_6 on KRAV_GRUNNLAG_431(VEDTAK_ID);
