+++ web/src/main/java/no/nav/foreldrepenger/tilbakekreving/web/app/tjenester/behandling/BehandlingFaktaRestTjeneste.java
@@ -4,8 +4,6 @@ import static javax.ws.rs.core.MediaType.APPLICATION_JSON;
 import static no.nav.vedtak.sikkerhet.abac.BeskyttetRessursActionAttributt.READ;
 import static no.nav.vedtak.sikkerhet.abac.BeskyttetRessursResourceAttributt.FAGSAK;
 
-import java.util.Optional;
-
 import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
 import javax.validation.Valid;
@@ -17,7 +15,7 @@ import javax.ws.rs.Produces;
 import javax.ws.rs.QueryParam;
 
 import io.swagger.annotations.Api;
-import no.nav.foreldrepenger.tilbakekreving.behandling.BehandlingTjeneste;
+import no.nav.foreldrepenger.tilbakekreving.behandling.impl.FaktaFeilutbetalingTjeneste;
 import no.nav.foreldrepenger.tilbakekreving.behandling.modell.BehandlingFeilutbetalingFakta;
 import no.nav.foreldrepenger.tilbakekreving.web.app.tjenester.behandling.dto.BehandlingFeilutbetalingFaktaDto;
 import no.nav.foreldrepenger.tilbakekreving.web.app.tjenester.behandling.dto.BehandlingIdDto;
@@ -34,26 +32,23 @@ public class BehandlingFaktaRestTjeneste {
 
     public static final String PATH_FRAGMENT = "/behandlingfakta";
 
-    private BehandlingTjeneste behandlingTjeneste;
+    private FaktaFeilutbetalingTjeneste faktaFeilutbetalingTjeneste;
 
     public BehandlingFaktaRestTjeneste() {
     }
 
     @Inject
-    public BehandlingFaktaRestTjeneste(BehandlingTjeneste behandlingTjeneste) {
-        this.behandlingTjeneste = behandlingTjeneste;
+    public BehandlingFaktaRestTjeneste(FaktaFeilutbetalingTjeneste faktaFeilutbetalingTjeneste) {
+        this.faktaFeilutbetalingTjeneste = faktaFeilutbetalingTjeneste;
     }
 
     @GET
     @Path("/hent-fakta/feilutbetaling")
     @BeskyttetRessurs(action = READ, ressurs = FAGSAK)
     public BehandlingFeilutbetalingFaktaDto hentFeilutbetalingFakta(@QueryParam(value = "behandlingId") @NotNull @Valid BehandlingIdDto idDto) {
-        BehandlingFeilutbetalingFaktaDto dto = null;
-        Optional<BehandlingFeilutbetalingFakta> behandlingFeilutbetalingFakta = behandlingTjeneste.hentBehandlingFeilutbetalingFakta(idDto.getBehandlingId());
-        if (behandlingFeilutbetalingFakta.isPresent()) {
-            dto = new BehandlingFeilutbetalingFaktaDto();
-            dto.setBehandlingFakta(behandlingFeilutbetalingFakta.get());
-        }
+        BehandlingFeilutbetalingFakta fakta = faktaFeilutbetalingTjeneste.hentBehandlingFeilutbetalingFakta(idDto.getBehandlingId());
+        BehandlingFeilutbetalingFaktaDto dto = new BehandlingFeilutbetalingFaktaDto();
+        dto.setBehandlingFakta(fakta);
         return dto;
     }
