+++ web/src/test/java/no/nav/foreldrepenger/tilbakekreving/web/app/tjenester/behandling/BehandlingFaktaRestTjenesteTest.java
@@ -8,11 +8,10 @@ import static org.mockito.Mockito.when;
 import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.Collections;
-import java.util.Optional;
 
 import org.junit.Test;
 
-import no.nav.foreldrepenger.tilbakekreving.behandling.BehandlingTjeneste;
+import no.nav.foreldrepenger.tilbakekreving.behandling.impl.FaktaFeilutbetalingTjeneste;
 import no.nav.foreldrepenger.tilbakekreving.behandling.modell.BehandlingFeilutbetalingFakta;
 import no.nav.foreldrepenger.tilbakekreving.behandling.modell.UtbetaltPeriode;
 import no.nav.foreldrepenger.tilbakekreving.web.app.tjenester.behandling.dto.BehandlingFeilutbetalingFaktaDto;
@@ -20,13 +19,13 @@ import no.nav.foreldrepenger.tilbakekreving.web.app.tjenester.behandling.dto.Beh
 
 public class BehandlingFaktaRestTjenesteTest {
 
-    private BehandlingTjeneste behandlingTjeneste = mock(BehandlingTjeneste.class);
+    private FaktaFeilutbetalingTjeneste behandlingTjeneste = mock(FaktaFeilutbetalingTjeneste.class);
 
     private BehandlingFaktaRestTjeneste restTjeneste = new BehandlingFaktaRestTjeneste(behandlingTjeneste);
 
     @Test
     public void skalHenteFeilutbetalingFakta() {
-        when(behandlingTjeneste.hentBehandlingFeilutbetalingFakta(anyLong())).thenReturn(Optional.of(lagBehandlingFeilutbetalingFakta()));
+        when(behandlingTjeneste.hentBehandlingFeilutbetalingFakta(anyLong())).thenReturn(lagBehandlingFeilutbetalingFakta());
         BehandlingIdDto dto = new BehandlingIdDto(123455L);
 
         BehandlingFeilutbetalingFaktaDto resultat = restTjeneste.hentFeilutbetalingFakta(dto);
@@ -36,13 +35,13 @@ public class BehandlingFaktaRestTjenesteTest {
 
     private BehandlingFeilutbetalingFakta lagBehandlingFeilutbetalingFakta() {
         return BehandlingFeilutbetalingFakta.builder()
-                .medAktuellFeilUtbetaltBeløp(BigDecimal.valueOf(2500))
-                .medDatoForRevurderingsvedtak(LocalDate.now().minusDays(40))
-                .medPerioder(Collections.singletonList(UtbetaltPeriode.lagPeriode(LocalDate.now().minusDays(120), LocalDate.now().minusDays(80), BigDecimal.valueOf(20000))))
-                .medTidligereVarsletBeløp(3000l)
-                .medTotalPeriodeFom(LocalDate.now().minusDays(150))
-                .medTotalPeriodeTom(LocalDate.now().minusDays(70))
-                .build();
+            .medAktuellFeilUtbetaltBeløp(BigDecimal.valueOf(2500))
+            .medDatoForRevurderingsvedtak(LocalDate.now().minusDays(40))
+            .medPerioder(Collections.singletonList(UtbetaltPeriode.lagPeriode(LocalDate.now().minusDays(120), LocalDate.now().minusDays(80), BigDecimal.valueOf(20000))))
+            .medTidligereVarsletBeløp(3000l)
+            .medTotalPeriodeFom(LocalDate.now().minusDays(150))
+            .medTotalPeriodeTom(LocalDate.now().minusDays(70))
+            .build();
     }
 
 }
